fix(rust): resolve all Rust backend exceptions from prod exception tracker
Tier 1 — SQL type/enum hard errors (broken endpoints):
- watch_history.rs: cast action param as ::watchaction in dynamic WHERE
- user_library.rs: decode RETURNING id/media_id as i32 (INT4 not INT8)
- rss.rs: decode rss_feed.id as i32 in user_update_rss_feed check
- content/streams.rs: SELECT stream_type::text to decode streamtype enum
- content/user_metadata.rs: fix ::media_type_enum → ::mediatype; fix
  series_id/season_id/genre_id/media_id/creator_id to i32; add missing
  mediatype cast in list query too
- admin_scrapers.rs: remove invalid ON CONFLICT on UPDATE; guard with
  NOT EXISTS subquery; bind ids as i32

Tier 2 — Data integrity bugs & panic:
- db/streams.rs: supply required NOT-NULL cols in file_media_link INSERT
  (is_primary, confidence, link_source, created_at)
- ptt/engine.rs: use floor_char_boundary for all byte-index string slices
  to prevent scraper task panics on multibyte UTF-8 titles
- scrapers/persist.rs: strip NUL bytes (0x00) from usenet stream text
  fields before Postgres insert; add strip_nul() helper

Tier 3 — HTTP 414 / oversized GET requests:
- premiumize.rs: batch check_cached into chunks of 80 hashes per request
- stremthru.rs: batch check_cached into chunks of 80 hashes per request

Tier 4 — Log noise reduction:
- providers/mod.rs: treat reqwest decode errors as transient (→ debug),
  collapsing 138k playback.rs warnings
- torrents/torbox.rs: check HTTP status before JSON decode; downgrade
  checkcached decode failure to debug, collapsing 45k warnings
- routes/poster.rs: downgrade upstream 404 and annotate-failed to debug
- jobs/runner.rs: use PgListener::connect() (dedicated standalone conn)
  instead of connect_with(&pool) to prevent NotificationResponse on
  shared pool query connections

+++ backend/src/routes/user_library.rs
@@ -441,9 +441,9 @@ pub async fn add_to_library(
     };
 
     // Verify media exists
-    let media: Option<(i64, String)> =
+    let media: Option<(i32, String)> =
         match sqlx::query_as("SELECT id, title FROM media WHERE id = $1")
-            .bind(body.media_id)
+            .bind(body.media_id as i32)
             .fetch_optional(&state.pool)
             .await
         {
@@ -466,11 +466,11 @@ pub async fn add_to_library(
     };
 
     // Check for existing entry
-    let existing: Option<i64> = sqlx::query_scalar(
+    let existing: Option<i32> = sqlx::query_scalar(
         "SELECT id FROM user_library_item WHERE user_id = $1 AND media_id = $2 LIMIT 1",
     )
     .bind(user_id as i32)
-    .bind(body.media_id)
+    .bind(body.media_id as i32)
     .fetch_optional(&state.pool)
     .await
     .unwrap_or(None);
@@ -493,14 +493,14 @@ pub async fn add_to_library(
     .unwrap_or(None);
 
     // Insert
-    let row: (i64, i64, String, String, Option<String>, DateTime<Utc>) =
+    let row: (i32, i32, String, String, Option<String>, DateTime<Utc>) =
         match sqlx::query_as(
             r#"INSERT INTO user_library_item (user_id, media_id, catalog_type, title_cached, poster_cached, added_at)
                VALUES ($1, $2, $3, $4, $5, NOW())
                RETURNING id, media_id, catalog_type, title_cached, poster_cached, added_at"#,
         )
         .bind(user_id as i32)
-        .bind(body.media_id)
+        .bind(body.media_id as i32)
         .bind(&body.catalog_type)
         .bind(&media_title)
         .bind(&poster_cached)
@@ -516,8 +516,8 @@ pub async fn add_to_library(
 
     let ext = get_external_ids(&state.pool, body.media_id as i32).await;
     let item = build_library_item(
-        row.0,
-        row.1,
+        row.0 as i64,
+        row.1 as i64,
         &row.2,
         &row.3,
         row.4.as_deref(),
